terraform_code/backend/backend_api/api/requirements.txt
Flask==2.2.3
Flask_Bcrypt==1.0.1
Flask_Cors==3.0.10
flask_mongoengine==1.0.0
Flask_RESTful==0.3.9
pandas==1.5.3
PyGithub==1.58.0
PyGithub==1.58.2
